change2double

+++ Metoder/4Momsen/Program.cs
@@ -5,10 +5,10 @@
         int vissSumma = 1337;
         System.Console.WriteLine($"Momsen på {vissSumma} är {Momsen(vissSumma)}");
     }
-    public static decimal Momsen(int Summan)
+    public static double Momsen(int Summan)
     {
 
-        decimal momsen = Summan * 0.25m;
+        double momsen = Summan * 0.25;
         return momsen;
     }
 }
